+++ buildDoc.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -eu
+#!/bin/bash -eux
 # Generate OpenAPI documentation
 GOPATH=${GOPATH:-~/go}
 echo "Using GOPATH: ${GOPATH}"