+++ api/forgot_test.go
@@ -17,8 +17,87 @@ func TestForgotResponds(t *testing.T) {
 	tests := []toTest{
 		{
 			// always returns a 200 if properly formed
-			method:   "POST",
-			url:      "/send/forgot/me@myemail.com",
+			// no language header -> default to EN
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Password reset instructions",
+			respCode:     200,
+		},
+		{
+			// testing language preferences
+			// follow standard header -> EN
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Password reset instructions",
+			customHeaders: map[string]string{
+				"Accept-Language": "en",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// follow standard header -> FR
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Réinitialisation du mot de passe",
+			customHeaders: map[string]string{
+				"Accept-Language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// standard header not supported language -> EN
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Password reset instructions",
+			customHeaders: map[string]string{
+				"Accept-Language": "gr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// follow custom header -> FR
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Réinitialisation du mot de passe",
+			customHeaders: map[string]string{
+				"x-tidepool-language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header not supported language -> EN
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Password reset instructions",
+			customHeaders: map[string]string{
+				"x-tidepool-language": "gr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header takes precedence over standard -> FR
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Réinitialisation du mot de passe",
+			customHeaders: map[string]string{
+				"Accept-Language": "en", "x-tidepool-language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header takes precedence over standard but language does not exist -> EN
+			method:       "POST",
+			url:          "/send/forgot/me@myemail.com",
+			emailSubject: "Password reset instructions",
+			customHeaders: map[string]string{
+				"Accept-Language": "en", "x-tidepool-language": "gr",
+			},
 			respCode: 200,
 		},
 		{
@@ -144,6 +223,11 @@ func TestForgotResponds(t *testing.T) {
 		if test.token != "" {
 			request.Header.Set(TP_SESSION_TOKEN, testing_token)
 		}
+		if test.customHeaders != nil {
+			for header, value := range test.customHeaders {
+				request.Header.Set(header, value)
+			}
+		}
 		response := httptest.NewRecorder()
 		testRtr.ServeHTTP(response, request)
 
@@ -163,6 +247,13 @@ func TestForgotResponds(t *testing.T) {
 
 			if cmp := result.deepCompare(&test.response); cmp != "" {
 				t.Fatalf("Test %d url: '%s'\n\t%s\n", idx, test.url, cmp)
+			}
+		}
+
+		if test.emailSubject != "" {
+			if emailSubjectSent := mockNotifier.GetLastEmailSubject(); emailSubjectSent != test.emailSubject {
+				t.Fatalf("Test %d url: '%s'\nNon-expected email subject %s (expected %s)",
+					idx, test.url, emailSubjectSent, test.emailSubject)
 			}
 		}
 	}