+++ api/signup.go
@@ -163,10 +163,14 @@ func (a *Api) sendSignUpInformation(res http.ResponseWriter, req *http.Request, 
 // @Description  This post is sent by the signup logic. In this state, the user account has been created but has a flag that
 // @Description  forces the user to the confirmation-required page until the signup has been confirmed.
 // @Description  It sends an email that contains a random confirmation link
+// @Description  The email is sent in the language defined by "x-tidepool-language" header
+// @Description  Otherwise if null "Accept-Language" header otherwise if null "ENglish"
 // @ID hydrophone-api-sendSignUp
 // @Accept  json
 // @Produce  json
 // @Param userid path string true "user id"
+// @Param x-tidepool-language header string false "User chosen language on 2 characters"
+// @Param Accept-Language header string false "Browser defined languages as array of languages such as fr-FR"
 // @Success 200 {string} string "OK"
 // @Failure 400 {object} status.Status "userId was not provided, return:\"Required userid is missing\" "
 // @Failure 401 {object} status.Status "Authorization token is missing or does not provided sufficient privileges"
@@ -283,11 +287,15 @@ func (a *Api) sendSignUp(res http.ResponseWriter, req *http.Request, vars map[st
 						emailContent["CreatorName"] = newSignUp.Creator.Profile.FullName
 					}
 
-					// although technically there exists a profile at the signup stage, the preferred language would always be empty here
-					// as it is set in the app and once the signup procedure is complete (after signup email has been confirmed)
-					// -> get browser's or "en" for English in case there is no browser's
-					if signerLanguage = GetBrowserPreferredLanguage(req); signerLanguage == "" {
-						signerLanguage = "en"
+					// on the "signup" page in Blip, the preferred language is now selected by listbox
+					// even if not selected there is one by default so we should normally always end up with
+					// a value in GetUserChosenLanguage()
+					// however, just to play it safe, we can continue taking the browser preferred language
+					// or ENglish as default values
+					if signerLanguage = GetUserChosenLanguage(req); signerLanguage == "" {
+						if signerLanguage = GetBrowserPreferredLanguage(req); signerLanguage == "" {
+							signerLanguage = "en"
+						}
 					}
 
 					if a.createAndSendNotification(req, newSignUp, emailContent, signerLanguage) {
@@ -365,9 +373,12 @@ func (a *Api) resendSignUp(res http.ResponseWriter, req *http.Request, vars map[
 
 				// although technically there exists a profile at the signup stage, the preferred language would always be empty here
 				// as it is set in the app and once the signup procedure is complete (after signup email has been confirmed)
-				// -> get browser's or "en" for English in case there is no browser's
-				if signerLanguage = GetBrowserPreferredLanguage(req); signerLanguage == "" {
-					signerLanguage = "en"
+				// so we rely on the language chosen by the user on the page through the list box
+				// otherwise if null it will be browser preferences otherwise EN
+				if signerLanguage = GetUserChosenLanguage(req); signerLanguage == "" {
+					if signerLanguage = GetBrowserPreferredLanguage(req); signerLanguage == "" {
+						signerLanguage = "en"
+					}
 				}
 
 				if a.createAndSendNotification(req, found, emailContent, signerLanguage) {