+++ api/signup_test.go
@@ -24,11 +24,104 @@ func TestSignupResponds(t *testing.T) {
 		},
 		{
 			// first time you ask, it does it
-			returnNone: true,
-			method:     "POST",
-			url:        "/send/signup/NewUserID",
-			token:      testing_token_uid1,
-			respCode:   200,
+			// no language header -> default to EN
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/NewUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Verify your email address",
+			respCode:     200,
+		},
+		{
+			// testing language preferences
+			// follow standard header -> EN
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/EnglishUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"Accept-Language": "en",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// follow standard header -> FR
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/FrenchUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Vérification de votre adresse email",
+			customHeaders: map[string]string{
+				"Accept-Language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// standard header not supported language -> EN
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/FrenchUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"Accept-Language": "gr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// follow custom header -> FR
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/FrenchUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Vérification de votre adresse email",
+			customHeaders: map[string]string{
+				"x-tidepool-language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header not supported language -> EN
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/FrenchUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"x-tidepool-language": "gr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header takes precedence over standard -> FR
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/FrenchUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Vérification de votre adresse email",
+			customHeaders: map[string]string{
+				"Accept-Language": "en", "x-tidepool-language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header takes precedence over standard but language does not exist -> EN
+			returnNone:   true,
+			method:       "POST",
+			url:          "/send/signup/FrenchUserID",
+			token:        testing_token_uid1,
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"Accept-Language": "fr", "x-tidepool-language": "gr",
+			},
+			respCode: 200,
 		},
 		{
 			// need a token
@@ -52,8 +145,86 @@ func TestSignupResponds(t *testing.T) {
 		},
 		{
 			// no token is all good
-			method:   "POST",
-			url:      "/resend/signup/email.resend@address.org",
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Verify your email address",
+			respCode:     200,
+		},
+		{
+			// testing language preferences
+			// follow standard header -> EN
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"Accept-Language": "en",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// follow standard header -> FR
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Vérification de votre adresse email",
+			customHeaders: map[string]string{
+				"Accept-Language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// standard header not supported language -> EN
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"Accept-Language": "gr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// follow custom header -> FR
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Vérification de votre adresse email",
+			customHeaders: map[string]string{
+				"x-tidepool-language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header not supported language -> EN
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"x-tidepool-language": "gr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header takes precedence over standard -> FR
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Vérification de votre adresse email",
+			customHeaders: map[string]string{
+				"Accept-Language": "en", "x-tidepool-language": "fr",
+			},
+			respCode: 200,
+		},
+		{
+			// testing language preferences
+			// custom header takes precedence over standard but language does not exist -> EN
+			method:       "POST",
+			url:          "/resend/signup/email.resend@address.org",
+			emailSubject: "Verify your email address",
+			customHeaders: map[string]string{
+				"Accept-Language": "fr", "x-tidepool-language": "gr",
+			},
 			respCode: 200,
 		},
 		{
@@ -251,6 +422,11 @@ func TestSignupResponds(t *testing.T) {
 		if test.token != "" {
 			request.Header.Set(TP_SESSION_TOKEN, testing_token)
 		}
+		if test.customHeaders != nil {
+			for header, value := range test.customHeaders {
+				request.Header.Set(header, value)
+			}
+		}
 		response := httptest.NewRecorder()
 		testRtr.ServeHTTP(response, request)
 
@@ -270,6 +446,13 @@ func TestSignupResponds(t *testing.T) {
 
 			if cmp := result.deepCompare(&test.response); cmp != "" {
 				t.Fatalf("Test %d url: '%s'\n\t%s\n", idx, test.url, cmp)
+			}
+		}
+
+		if test.emailSubject != "" {
+			if emailSubjectSent := mockNotifier.GetLastEmailSubject(); emailSubjectSent != test.emailSubject {
+				t.Fatalf("Test %d url: '%s'\nNon-expected email subject %s (expected %s)",
+					idx, test.url, emailSubjectSent, test.emailSubject)
 			}
 		}
 	}